working copy of RGBW conversions in color1.py
